requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
graphql-core==3.2.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.270.2
strawberry-graphql-django==0.60.0
typing_extensions==4.13.2
uritemplate==4.1.1
